bs-templates/bs/{{cookiecutter.__project_slug}}/requirements.local.txt
Flask>=0.9
pandas>=0.23.0
numpy<1.24
python-dotenv==0.19.0
webaiku @ git+https://github.com/dataiku/solutions-contrib.git@feature/packaging#egg=webaiku&subdirectory=bs-infra
dataiku-api-client>=11.0.0